Batteries are now implemented
