+++ src/main/java/task_05_Adapter/Box.java
@@ -1,7 +1,7 @@
 package task_05_Adapter;
 
-import task_03_Composite.Battery;
 import FLF.Status;
+import task_03_Composite.Battery;
 
 public class Box {
     private Status batterystatus;
@@ -34,7 +34,7 @@ public class Box {
             double cap =battery.getCapacity();
             if(cap-amount>0) {
                 cap = cap-amount/4;
-                battery.takeOut(cap);
+                battery.takeOut((int)cap);
             } else {
                 cap = 0;
                 return false;
