--- test/build.gradle
@@ -1,22 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'org.example'
-version '1.0-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.rhq.misc:misc:0.1.3'
-    implementation 'org.objectweb.util.misc:ow-misc:1.0'
-    implementation 'com.sun.jini:sun-util:2.1'
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.2'
-}
-
-test {
-    useJUnitPlatform()
-}
